refactored

+++ src/main/java/by/belhostel/hostels/command/admin/DeleteCommand.java
@@ -25,41 +25,48 @@ import java.util.List;
  */
 public class DeleteCommand implements ActionCommand {
     private static final Logger LOG = LogManager.getLogger();
+    private static final String PARAM_CURRENT_USER = "currentUser";
     private static final String PARAM_CLAIM_ID = "claimId";
     private static final String PARAM_MESSAGES = "messages";
     private static final String PARAM_UNCONFIRMED_CLAIMS = "unconfirmedClaims";
     private static final String PARAM_CLAIMS_PAGE = "/service?command=go&page=claims";
     private static final String PARAM_ERROR_MESSAGE = "errorMessage";
     private static final String PARAM_ERROR = "/resources/jsp/error.jsp";
+    private static final String PARAM_MAIN = "/resources/jsp/main.jsp";
 
     public String execute(HttpServletRequest request, HttpServletResponse response) {
         String page = "";
         HttpSession session = request.getSession(true);
-        try {
-            int claimId = Integer.parseInt(request.getParameter(PARAM_CLAIM_ID));
-            Claim claim = HostelManager.findClaimById(claimId);
-            Hostel current = HostelManager.findHostelById(claim.getHostelId());
-            User user = HostelManager.findUserById(claim.getUserId());
+        User currentUser = (User) session.getAttribute(PARAM_CURRENT_USER);
+        if (currentUser != null && currentUser.isAdmin()) {
+            try {
+                int claimId = Integer.parseInt(request.getParameter(PARAM_CLAIM_ID));
+                Claim claim = HostelManager.findClaimById(claimId);
+                Hostel current = HostelManager.findHostelById(claim.getHostelId());
+                User user = HostelManager.findUserById(claim.getUserId());
 
-            boolean res = AdminAction.deleteClaimById(claimId);
-            if (res) {
-                ArrayList<Claim> unconfirmedClaims = (ArrayList<Claim>) session.getAttribute(PARAM_UNCONFIRMED_CLAIMS);
-                unconfirmedClaims.remove(claim);
-                List<HttpSession> userSessions = UserSessions.getUserSessions(user.getUserId());
-                Message rejectionMessage = Messenger.generateRejectionMessage(user.getUserId(), current.getName(), claim);
-                boolean isSent = HostelManager.sendMessageToUser(rejectionMessage);
-                if (isSent) {
-                    for (HttpSession userSession : userSessions) {
-                        ArrayList<Message> messages = (ArrayList<Message>) userSession.getAttribute(PARAM_MESSAGES);
-                        messages.add(rejectionMessage);
+                boolean res = AdminAction.deleteClaimById(claimId);
+                if (res) {
+                    ArrayList<Claim> unconfirmedClaims = (ArrayList<Claim>) session.getAttribute(PARAM_UNCONFIRMED_CLAIMS);
+                    unconfirmedClaims.remove(claim);
+                    List<HttpSession> userSessions = UserSessions.getUserSessions(user.getUserId());
+                    Message rejectionMessage = Messenger.generateRejectionMessage(user.getUserId(), current.getName(), claim);
+                    boolean isSent = HostelManager.sendMessageToUser(rejectionMessage);
+                    if (isSent) {
+                        for (HttpSession userSession : userSessions) {
+                            ArrayList<Message> messages = (ArrayList<Message>) userSession.getAttribute(PARAM_MESSAGES);
+                            messages.add(rejectionMessage);
+                        }
                     }
                 }
+                response.sendRedirect(request.getContextPath() + PARAM_CLAIMS_PAGE);
+            } catch (IOException | NumberFormatException | LogicException e) {
+                LOG.error(e);
+                request.setAttribute(PARAM_ERROR_MESSAGE, e);
+                page = PARAM_ERROR;
             }
-            response.sendRedirect(request.getContextPath() + PARAM_CLAIMS_PAGE);
-        } catch (IOException | NumberFormatException | LogicException e) {
-            LOG.error(e);
-            request.setAttribute(PARAM_ERROR_MESSAGE, e);
-            page = PARAM_ERROR;
+        } else {
+            page = PARAM_MAIN;
         }
         return page;
     }
